requirement.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
idna==2.10
Pillow==8.1.0
pkg-resources==0.0.0
pymongo==3.11.2
requests==2.25.0cle
simplejson==3.17.2
soupsieve==2.1
urllib3==1.26.2
